move minimap into parent minimap component
